LayoutTests/media/audio-constructor-expected.txt

Test that Audio() object loads the resource after src attribute is set and load() is called.

TEST(audio instanceof HTMLAudioElement) OK
RUN(audio.load())
EVENT(loadstart) TEST(relativeURL(audio.currentSrc)=='content/test.wav') OK
EVENT(load)
END OF TEST
